requirements.txt
numpy
packaging
pillow
termcolor
matplotlib
cupy-cuda115>=10.0
